+++ examples/cli-chat/src/chat.ts
@@ -104,7 +104,9 @@ export default async function startChat(): Promise<void> {
     rl.prompt();
     const chatMessage = ChatMessage.fromUtf8String(new Date(), nick, line);
 
-    const msg = WakuMessage.fromBytes(chatMessage.encode(), ChatContentTopic);
+    const msg = WakuMessage.fromBytes(chatMessage.encode(), {
+      contentTopic: ChatContentTopic,
+    });
     if (opts.lightPush) {
       await waku.lightPush.push(msg);
     } else {